Updated the main file to analyze Clarifai output.
Takes a command line argument towards an image url, and creates a text file to store the tags found from Clarifai. This requires there to be a file without and extension labeled 'config' in the same directory as the main 'Food_Analyzer.py' file. The details for the config file are within the main python file near the top.