Added functionality
Created several functions so that one use an image database (with a root folder, an subfolders with the food name, and pictures inside of the food) to create multiple files with all tags found for the images. The files are labeled 'Tag_food_item.txt' and they are stored inside the 'Image_Tags' folder.